requirements.txt
aiohttp==3.6.2
asn1crypto==1.3.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
discord.py==1.3.1
humanize==0.5.1
idna==2.9
multidict==4.7.4
pycparser==2.19
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dotenv==0.11.0
requests==2.22.0
six==1.14.0
urllib3==1.25.8
websockets==8.1
win-inet-pton==1.1.0
wincertstore==0.2
yarl==1.4.2
